get tests running. still need xvfb for chrome

+++ build.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
 
-vagrant plugin install vagrant-vbguest
+if [[ ! -e .vbguest-installed ]]; then
+  touch .vbguest-installed
+  vagrant plugin install vagrant-vbguest
+fi
 vagrant up
 vagrant ssh <<-SHELL
   cd /home/vagrant/bookhost
-  ./gradlew clean war test
+  ./gradlew clean war
+  sudo cp build/libs/bookhost.war /opt/tomee/webapps
+  sudo chown tomee:tomee /opt/tomee/webapps/bookhost.war
   sudo cp build/libs/bookhost.war /build-output
   sudo systemctl restart tomee
 SHELL
